OrderGenerator now creates a string requesting six toppings.
OrderGenerator now chooses (and toggles) six toppings to be used in the
next order of the game. Adds all topping names to a string to be used
in the UI. Also contains “OrderReset” function to move order values to
default if initial order generation fails or when the round ends.